+++ lib/presentation/screens/counter/widgets/progress_header.dart
@@ -36,7 +36,7 @@ class ProgressHeader extends StatelessWidget {
           ? () => onLongPress!(context.getWidgetRect())
           : null,
       child: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 16),
+        padding: const EdgeInsets.only(left: 24, right: 24, top: 16, bottom: 8),
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [